ai/requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
apscheduler==3.10.4
pydantic==2.9.2
PyYAML==6.0.2
